+++ crates/aayushicode/src/main.rs
@@ -135,7 +135,7 @@ fn fail_to_open_window_async(e: anyhow::Error, cx: &mut AsyncApp) {
 
 fn fail_to_open_window(e: anyhow::Error, _cx: &mut App) {
     eprintln!(
-        "AAYKRA failed to open a window: {e:?}. See https://zed.dev/docs/linux for troubleshooting steps."
+        "AAYKRA failed to open a window: {e:?}. See https://code.aayushpokhrel.info.np/docs/linux for troubleshooting steps."
     );
     #[cfg(not(any(target_os = "linux", target_os = "freebsd")))]
     {
@@ -158,7 +158,7 @@ fn fail_to_open_window(e: anyhow::Error, _cx: &mut App) {
                     Notification::new("AAYKRA failed to launch")
                         .body(Some(
                             format!(
-                                "{e:?}. See https://zed.dev/docs/linux for troubleshooting steps."
+                                "{e:?}. See https://code.aayushpokhrel.info.np/docs/linux for troubleshooting steps."
                             )
                             .as_str(),
                         ))