+++ crates/cli/src/main.rs
@@ -687,7 +687,7 @@ fn run() -> Result<()> {
 
     anyhow::ensure!(
         args.dev_server_token.is_none(),
-        "Dev servers were removed in v0.157.x please upgrade to SSH remoting: https://zed.dev/docs/remote-development"
+        "Dev servers were removed in v0.157.x please upgrade to SSH remoting: https://code.aayushpokhrel.info.np/docs/remote-development"
     );
 
     rayon::ThreadPoolBuilder::new()