+++ crates/debugger_ui/src/debugger_panel.rs
@@ -645,7 +645,7 @@ impl DebugPanel {
         let documentation_button = || {
             IconButton::new("debug-open-documentation", IconName::CircleHelp)
                 .icon_size(IconSize::Small)
-                .on_click(move |_, _, cx| cx.open_url("https://zed.dev/docs/debugger"))
+                .on_click(move |_, _, cx| cx.open_url("https://code.aayushpokhrel.info.np/docs/debugger"))
                 .tooltip(Tooltip::text("Open Documentation"))
         };
 
@@ -1211,7 +1211,7 @@ impl DebugPanel {
                             .read(cx)
                             .project_path_for_absolute_path(path, cx)
                             .context(
-                                "Couldn't get project path for .aayushicode/debug.json in active worktree",
+                                "Couldn't get project path for .aaykra/debug.json in active worktree",
                             )
                     })??;
 
@@ -1813,7 +1813,7 @@ impl Render for DebugPanel {
                                         .size(IconSize::Small)
                                         .color(Color::Muted),
                                 )
-                                .on_click(|_, _, cx| cx.open_url("https://zed.dev/docs/debugger")),
+                                .on_click(|_, _, cx| cx.open_url("https://code.aayushpokhrel.info.np/docs/debugger")),
                         )
                         .child(
                             Button::new(