+++ crates/project/src/project.rs
@@ -2903,7 +2903,7 @@ impl Project {
                         notification_id: format!("local-tasks-{path:?}").into(),
                         link: Some(ToastLink {
                             label: "Open Tasks Documentation",
-                            url: "https://zed.dev/docs/tasks",
+                            url: "https://code.aayushpokhrel.info.np/docs/tasks",
                         }),
                         message,
                     });