+++ crates/remote/src/transport.rs
@@ -32,7 +32,7 @@ fn parse_platform(output: &str) -> Result<RemotePlatform> {
         "Darwin" => RemoteOs::MacOs,
         "Linux" => RemoteOs::Linux,
         _ => anyhow::bail!(
-            "Prebuilt remote servers are not yet available for {os:?}. See https://zed.dev/docs/remote-development"
+            "Prebuilt remote servers are not yet available for {os:?}. See https://code.aayushpokhrel.info.np/docs/remote-development"
         ),
     };
 
@@ -47,7 +47,7 @@ fn parse_platform(output: &str) -> Result<RemotePlatform> {
         RemoteArch::X86_64
     } else {
         anyhow::bail!(
-            "Prebuilt remote servers are not yet available for {arch:?}. See https://zed.dev/docs/remote-development"
+            "Prebuilt remote servers are not yet available for {arch:?}. See https://code.aayushpokhrel.info.np/docs/remote-development"
         )
     };