+++ crates/remote_server/src/remote_editing_tests.rs
@@ -703,7 +703,7 @@ async fn test_remote_settings(cx: &mut TestAppContext, server_cx: &mut TestAppCo
     });
 
     fs.insert_tree(
-        "/code/project1/.zed",
+        "/code/project1/.aayushicode",
         json!({
             "settings.json": r#"
                   {
@@ -790,7 +790,7 @@ async fn test_remote_lsp(cx: &mut TestAppContext, server_cx: &mut TestAppContext
     let (project, headless) = init_test(&fs, cx, server_cx).await;
 
     fs.insert_tree(
-        path!("/code/project1/.zed"),
+        path!("/code/project1/.aayushicode"),
         json!({
             "settings.json": r#"
           {
@@ -1029,7 +1029,7 @@ async fn test_remote_call_hierarchy(cx: &mut TestAppContext, server_cx: &mut Tes
     let (project, headless) = init_test(&fs, cx, server_cx).await;
 
     fs.insert_tree(
-        path!("/code/project1/.zed"),
+        path!("/code/project1/.aayushicode"),
         json!({
             "settings.json": r#"
           {
@@ -2127,7 +2127,7 @@ async fn test_remote_cancel_language_server_work(
     let (project, headless) = init_test(&fs, cx, server_cx).await;
 
     fs.insert_tree(
-        path!("/code/project1/.zed"),
+        path!("/code/project1/.aayushicode"),
         json!({
             "settings.json": r#"
           {
@@ -3881,7 +3881,7 @@ async fn test_remote_apply_code_action_skips_unadvertised_command(
     let (project, headless) = init_test(&fs, cx, server_cx).await;
 
     fs.insert_tree(
-        path!("/code/project1/.zed"),
+        path!("/code/project1/.aayushicode"),
         json!({
             "settings.json": r#"
           {