+++ crates/tasks_ui/src/tasks_ui.rs
@@ -491,7 +491,7 @@ mod tests {
         fs.insert_tree(
             path!("/dir"),
             json!({
-                ".zed": {
+                ".aayushicode": {
                     "tasks.json": r#"[
                             {
                                 "label": "example task",