+++ crates/theme_selector/src/icon_theme_selector.rs
@@ -333,7 +333,7 @@ impl PickerDelegate for IconThemeSelectorDelegate {
                                 .color(Color::Muted),
                         )
                         .on_click(|_event, _window, cx| {
-                            cx.open_url("https://zed.dev/docs/icon-themes");
+                            cx.open_url("https://code.aayushpokhrel.info.np/docs/icon-themes");
                         }),
                 )
                 .child(