+++ crates/workspace/src/notifications.rs
@@ -1307,7 +1307,7 @@ pub mod simple_message_notification {
                 fn primary_action(&self) -> ErrorAction {
                     ErrorAction::link(
                         "See Docs",
-                        "https://zed.dev/docs/linux#i-cant-open-any-files",
+                        "https://code.aayushpokhrel.info.np/docs/linux#i-cant-open-any-files",
                     )
                 }
             }