+++ crates/workspace/src/workspace_error.rs
@@ -198,7 +198,7 @@ impl WorkspaceError for PortalError {
     fn primary_action(&self) -> ErrorAction {
         ErrorAction::link(
             "See docs",
-            "https://zed.dev/docs/linux#i-cant-open-any-files",
+            "https://code.aayushpokhrel.info.np/docs/linux#i-cant-open-any-files",
         )
     }
 }